+++ src/main/java/com/cellterminal/network/PacketExtractUpgrade.java
@@ -8,8 +8,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
-import com.cellterminal.container.ContainerCellTerminal;
-import com.cellterminal.container.ContainerWirelessCellTerminal;
+import com.cellterminal.container.ContainerCellTerminalBase;
 
 
 /**
@@ -114,19 +113,9 @@ public class PacketExtractUpgrade implements IMessage {
             player.getServerWorld().addScheduledTask(() -> {
                 Container container = player.openContainer;
 
-                if (container instanceof ContainerCellTerminal) {
-                    ContainerCellTerminal cellContainer = (ContainerCellTerminal) container;
+                if (container instanceof ContainerCellTerminalBase) {
+                    ContainerCellTerminalBase cellContainer = (ContainerCellTerminalBase) container;
                     cellContainer.handleExtractUpgrade(
-                        player,
-                        message.getTargetType(),
-                        message.getTargetId(),
-                        message.getCellSlot(),
-                        message.getUpgradeIndex(),
-                        message.isToInventory()
-                    );
-                } else if (container instanceof ContainerWirelessCellTerminal) {
-                    ContainerWirelessCellTerminal wirelessContainer = (ContainerWirelessCellTerminal) container;
-                    wirelessContainer.handleExtractUpgrade(
                         player,
                         message.getTargetType(),
                         message.getTargetId(),