+++ src/main/java/com/cellterminal/network/PacketInsertCell.java
@@ -8,8 +8,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
-import com.cellterminal.container.ContainerCellTerminal;
-import com.cellterminal.container.ContainerWirelessCellTerminal;
+import com.cellterminal.container.ContainerCellTerminalBase;
 
 
 /**
@@ -49,12 +48,9 @@ public class PacketInsertCell implements IMessage {
             player.getServerWorld().addScheduledTask(() -> {
                 Container container = player.openContainer;
 
-                if (container instanceof ContainerCellTerminal) {
-                    ContainerCellTerminal cellContainer = (ContainerCellTerminal) container;
+                if (container instanceof ContainerCellTerminalBase) {
+                    ContainerCellTerminalBase cellContainer = (ContainerCellTerminalBase) container;
                     cellContainer.handleInsertCell(message.storageId, message.targetSlot, player);
-                } else if (container instanceof ContainerWirelessCellTerminal) {
-                    ContainerWirelessCellTerminal wirelessContainer = (ContainerWirelessCellTerminal) container;
-                    wirelessContainer.handleInsertCell(message.storageId, message.targetSlot, player);
                 }
             });