+++ src/main/java/com/cellterminal/network/PacketPartitionAction.java
@@ -11,8 +11,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
-import com.cellterminal.container.ContainerCellTerminal;
-import com.cellterminal.container.ContainerWirelessCellTerminal;
+import com.cellterminal.container.ContainerCellTerminalBase;
 
 
 /**
@@ -111,18 +110,9 @@ public class PacketPartitionAction implements IMessage {
             player.getServerWorld().addScheduledTask(() -> {
                 Container container = player.openContainer;
 
-                if (container instanceof ContainerCellTerminal) {
-                    ContainerCellTerminal cellContainer = (ContainerCellTerminal) container;
+                if (container instanceof ContainerCellTerminalBase) {
+                    ContainerCellTerminalBase cellContainer = (ContainerCellTerminalBase) container;
                     cellContainer.handlePartitionAction(
-                        message.storageId,
-                        message.cellSlot,
-                        message.action,
-                        message.partitionSlot,
-                        message.itemStack
-                    );
-                } else if (container instanceof ContainerWirelessCellTerminal) {
-                    ContainerWirelessCellTerminal wirelessContainer = (ContainerWirelessCellTerminal) container;
-                    wirelessContainer.handlePartitionAction(
                         message.storageId,
                         message.cellSlot,
                         message.action,