+++ src/main/java/com/cellterminal/network/PacketUpgradeCell.java
@@ -11,8 +11,7 @@ import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessageHandler;
 import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 
-import com.cellterminal.container.ContainerCellTerminal;
-import com.cellterminal.container.ContainerWirelessCellTerminal;
+import com.cellterminal.container.ContainerCellTerminalBase;
 
 
 /**
@@ -95,18 +94,9 @@ public class PacketUpgradeCell implements IMessage {
             player.getServerWorld().addScheduledTask(() -> {
                 Container container = player.openContainer;
 
-                if (container instanceof ContainerCellTerminal) {
-                    ContainerCellTerminal cellContainer = (ContainerCellTerminal) container;
+                if (container instanceof ContainerCellTerminalBase) {
+                    ContainerCellTerminalBase cellContainer = (ContainerCellTerminalBase) container;
                     cellContainer.handleUpgradeCell(
-                        player,
-                        message.getStorageId(),
-                        message.getCellSlot(),
-                        message.isShiftClick(),
-                        message.getFromSlot()
-                    );
-                } else if (container instanceof ContainerWirelessCellTerminal) {
-                    ContainerWirelessCellTerminal wirelessContainer = (ContainerWirelessCellTerminal) container;
-                    wirelessContainer.handleUpgradeCell(
                         player,
                         message.getStorageId(),
                         message.getCellSlot(),